gen-data/requirements.txt
psycopg2-binary
python-dotenv
faker